Лекция 8.3 и Лекция 8.4

+++ mantis-tests/src/test/java/itsm/mantis/appmanager/ApplicationManager.java
@@ -9,11 +9,13 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.IOException;
 import java.util.Properties;
+import java.util.concurrent.TimeUnit;
 
 public class ApplicationManager {
     private final String browser;
     private final Properties properties;
     private WebDriver wd;
+    RegistrationHelper registrationHelper;
 
     public ApplicationManager(String browser) {
         this.browser = browser;
@@ -25,20 +27,43 @@ public class ApplicationManager {
         String driver = System.getProperty("driver", "browsers");
         properties.load(new FileReader(new File(String.format("src/test/resources/%s.properties", driver))));
         properties.load(new FileReader(new File(String.format("src/test/resources/%s.properties", target))));
-        if (browser.equals(BrowserType.CHROME)) {
-            System.setProperty(properties.getProperty("web.chromeDriver"), properties.getProperty("web.forChrome"));
-            wd = new ChromeDriver();
-        } else if (browser.equals(BrowserType.FIREFOX)) {
-            System.setProperty(properties.getProperty("web.firefoxDriver"), properties.getProperty("web.forFirefox"));
-            wd = new FirefoxDriver();
-        }
-        wd.manage().window().maximize();
-        wd.manage().deleteAllCookies();
-//        wd.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
-        wd.get(properties.getProperty("web.baseUrl"));
     }
 
     public void stop() {
-        wd.close();
+        if(wd != null) {
+            wd.close();
+        }
+    }
+
+    public HttpSession newSession(){
+        return new HttpSession(this);
+    }
+
+    public String getProperty(String key){
+        return properties.getProperty(key);
+    }
+
+    public RegistrationHelper registration() {
+        if(registrationHelper == null) {
+            registrationHelper = new RegistrationHelper(this);
+        }
+        return registrationHelper;
+    }
+
+    public WebDriver getDriver() {
+        if(wd == null){
+            if (browser.equals(BrowserType.CHROME)) {
+                System.setProperty(properties.getProperty("web.chromeDriver"), properties.getProperty("web.forChrome"));
+                wd = new ChromeDriver();
+            } else if (browser.equals(BrowserType.FIREFOX)) {
+                System.setProperty(properties.getProperty("web.firefoxDriver"), properties.getProperty("web.forFirefox"));
+                wd = new FirefoxDriver();
+            }
+            wd.manage().window().maximize();
+            wd.manage().deleteAllCookies();
+            wd.manage().timeouts().implicitlyWait(30, TimeUnit.SECONDS);
+            wd.get(properties.getProperty("web.baseUrl"));
+        }
+        return wd;
     }
 }
